fix: pending migrations in migrate mongodb gh action

+++ src/migrations/2022121620560802-remove-phone-code.ts
@@ -1,7 +1,11 @@
 module.exports = {
   async up(db) {
-    console.log("removing users")
-    const updated = await db.collection("phonecodes").drop()
-    console.log({ updated }, "updated users")
+    console.log("removing phonecodes")
+    const collectionList = await db.listCollections({ name: "phonecodes" }).toArray()
+    if (collectionList.length > 0) {
+      const updated = await db.collection("phonecodes").drop()
+      console.log({ updated }, "phonecodes removed")
+    }
+    console.log("phonecodes already removed")
   },
 }
